requirements.txt
langchain-gigachat>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn>=0.32.0
python-telegram-bot>=21.0
httpx>=0.27.0
ddgs>=8.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-timeout>=2.3.1
